feat: Add funtionality to send assessment
