+++ src/Pages/ViewAssessmentDraft.js
@@ -1,5 +1,7 @@
 import PropTypes from 'prop-types';
 import { Slider } from '@mui/material';
+import { useEffect, useState } from "react";
+import emailjs from "@emailjs/browser";
 
 const ViewAssessmentDraft = ({ assessmentQuestions, deleteQuestion }) => {
   const marks = [
@@ -47,10 +49,57 @@ const ViewAssessmentDraft = ({ assessmentQuestions, deleteQuestion }) => {
     </div>
   ));
 
+  const [loading, setLoading] = useState(false)
+
+
+  useEffect(() => emailjs.init("zo743h247QwTwz8gU"), []);
+  //... 
+    const [recipientName, setRecipientName] = useState('');
+    const [recipientEmail, setRecipientEmail] = useState('');
+
+    const handleSubmit = async (e) => {
+    e.preventDefault();
+    const serviceId = "service_qzyuot1";
+    const templateId = "template_7zacm7f";
+    try {
+      setLoading(true);
+      await emailjs.send(serviceId, templateId, {
+       name: recipientName,
+       recipient: recipientEmail
+      });
+      alert("email successfully sent check inbox");
+    } catch (error) {
+      console.log(error);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+
   return (
     <>
       <p>ViewAssessmentDraft</p>
       {displayAssessmentQuestions}
+
+      <aside></aside>
+      <form className="for" onSubmit={handleSubmit}>
+          <h5> Who would you like to send the assessment to?</h5>
+          <label htmlFor="name" aria-label="name">name*</label>
+          <input
+            type="text"
+            placeholder="Enter recipient name"
+            value={recipientName}
+            onChange={(e) => setRecipientName(e.target.value)}
+          />
+          <label htmlFor="email" aria-label="email">email*</label>
+          <input
+            type="email"
+            placeholder="Enter recipient email"
+            value={recipientEmail}
+            onChange={(e) => setRecipientEmail(e.target.value)}
+        />
+          <button type="submit" disabled={loading}  >Publish assessment</button>
+      </form>
     </>
   );
 };
